+++ src/Blur.cpp
@@ -1,4 +1,4 @@
-#ifdef GEODE_IS_DESKTOP
+/*#ifdef GEODE_IS_DESKTOP
 
 #include "Blur.hpp"
 
@@ -235,12 +235,11 @@ void cleanupPostProcess() {
 
 class $modify(CCNode) {
     void visit() {
-        if ((CCNode*)this != CCDirector::get()->getRunningScene() || ppShader.program == 0) {
-            CCNode::visit();
-            return;
-        }
+        if (this->getID() == std::string("android-ui"))
+            blurTimer = as<float>(as<CCLayerColor*>((CCNode*)(this))->getOpacity()) / 100.0f;
+        
         float blur = 0.05 * (1.f - std::cos((float)std::numbers::pi * blurTimer)) * 0.5f;
-        if (blur == 0.f) {
+        if (blur == 0.f || /*this->getID() != std::string("android-ui") || (CCNode*)this != CCDirector::get()->getRunningScene() || ppShader.program == 0) {
             CCNode::visit();
             return;
         }
@@ -318,7 +317,7 @@ $on_mod(Unloaded) {
     cleanupPostProcess();
 }
 
-class $modify(CCScheduler) {
+/*class $modify(CCScheduler) {
     void update(float dt) {
         CCScheduler::update(dt);
         
@@ -329,4 +328,4 @@ class $modify(CCScheduler) {
     }
 };
 
-#endif
+#endif*/