+++ src/Hacks/HitboxTrail.cpp
@@ -4,34 +4,40 @@
 
 using namespace geode::prelude;
 
+std::vector<CCPoint> points;
+
 class $modify (GJBaseGameLayer)
 {
+    CCPoint lastPos = ccp(0, 0);
+
     Module* mod = nullptr;
     Module* mod2 = nullptr;
 
     CCDrawNode* dn = nullptr;
 
-    std::vector<CCPoint> points = {};
-
     virtual TodoReturn updateDebugDraw()
     {
         GJBaseGameLayer::updateDebugDraw();
 
-        if (!m_fields->mod)
-            m_fields->mod = Client::GetModule("hitbox-trail");
+        //if (!m_fields->mod)
+            //m_fields->mod = Client::GetModule("hitbox-trail");
 
-        if (m_fields->mod->enabled && m_player1)
+        if (/*m_fields->mod->enabled*//*true && m_player1 && m_debugDrawNode)
         {
-            if (m_fields->points.size() == 0 || (m_fields->points[m_fields->points.size() - 1].equals(m_player1->getPosition())))
+            if (m_fields->lastPos != m_player1->getPosition())
             {
-                m_fields->points.push_back(m_player1->getPosition());
+                m_fields->lastPos = m_player1->getPosition();
+
+                points.push_back(m_player1->getPosition());
             }
 
-            for (size_t i = 0; i < m_fields->points.size(); i++)
+            if (points.size() != 0)
             {
-                m_debugDrawNode->drawRect(m_fields->points[i], m_fields->points[i] + ccp(100, 100), ccc4FFromccc4B(ccc4(0, 0, 0, 0)), 1, ccc4FFromccc4B(ccc4(255, 0, 0, 255)));
-            }
-            
+                for (size_t i = 0; i < points.size(); i++)
+                {
+                    m_debugDrawNode->drawRect(points[i], points[i] + ccp(100, 100), ccc4FFromccc4B(ccc4(0, 0, 0, 0)), 1, ccc4FFromccc4B(ccc4(255, 0, 0, 255)));
+                }
+            }            
         }
     }
 };*/